add filter and sort, Add analysis type 2, add map visualization, create update logic form visualizations
created filter and sort for each column on text viz

created analysis type 2

added plotly for map vizualization inside an iframe. Also had to fork and fix library because the last commit to go-plotly broke it.

added last analysis type field to keep track of when the analysis type changes.